+++ packages/core/src/server/index.ts
@@ -4,9 +4,3 @@ export * as PageTree from './page-tree';
 export * from './git-api';
 export * from './types';
 export * from './metadata';
-
-// TODO: Remove on next major
-/**
- * @deprecated
- */
-export * from '../highlight';