+++ packages/create-app/src/versions.d.ts
@@ -2,5 +2,6 @@ export const versions: {
   'fumadocs-core': string;
   'fumadocs-ui': string;
   'fumadocs-mdx': string;
+  '@fumadocs/mdx-remote': string;
   '@fumadocs/content-collections': string;
 };