+++ packages/mdx-remote/tsup.config.ts
@@ -4,6 +4,6 @@ export default defineConfig({
   external: ['fumadocs-core', 'next', 'react'],
   dts: true,
   target: 'es2021',
-  entry: ['./src/index.ts'],
+  entry: ['./src/index.ts', './src/client/index.ts'],
   format: 'esm',
 });